requirements.txt
opshin==0.24.0
Flask~=3.1
python-dotenv~=1.0
bcrypt~=4.2
python-ulid~=3.0
Flask-SQLAlchemy~=3.1